backend/requirements.txt
Django==4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
mysqlclient==2.2.0
